refactor: migrate color variants to a new colors module and update imports
